use FontMetricsContext to consume/provide font metrics
