+++ demo/src/solver/substeps.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import * as Editor from "@math-blocks/editor-core";
-import fontMetrics from "@math-blocks/metrics";
+import {FontMetricsContext} from "@math-blocks/metrics";
 import {MathRenderer} from "@math-blocks/react";
 import {types} from "@math-blocks/semantic";
 import {Step, applyStep} from "@math-blocks/step-utils";
@@ -18,6 +18,8 @@ type Props = {
 };
 
 const Substeps: React.FunctionComponent<Props> = ({prefix, start, step}) => {
+    const fontMetrics = React.useContext(FontMetricsContext);
+
     let current = start;
 
     const fontSize = 64;
