+++ demo/src/tutor/tutor-page.tsx
@@ -1,6 +1,8 @@
 import * as React from "react";
 import {Provider} from "react-redux";
 
+import {comicSans, FontMetricsContext} from "@math-blocks/metrics";
+
 import {store} from "./store";
 import Tutor from "./tutor";
 
@@ -10,7 +12,9 @@ console.log(store.getState());
 const TutorPage: React.FunctionComponent = () => {
     return (
         <Provider store={store}>
-            <Tutor />
+            <FontMetricsContext.Provider value={comicSans}>
+                <Tutor />
+            </FontMetricsContext.Provider>
         </Provider>
     );
 };
