--- packages/metrics/src/comics-sans.ts
@@ -1,4057 +1,0 @@
-import type {FontMetrics} from "./types";
-
-const metrics: FontMetrics = {
-    unitsPerEm: 2048,
-    ascender: 2257,
-    descender: -597,
-    glyphMetrics: {
-        "32": {
-            advance: 612,
-            bearingX: 0,
-            bearingY: 0,
-            width: 0,
-            height: 0,
-        },
-        "33": {
-            advance: 487,
-            bearingX: 120,
-            bearingY: 1605,
-            width: 212,
-            height: 1671,
-        },
-        "34": {
-            advance: 869,
-            bearingX: 116,
-            bearingY: 1588,
-            width: 571,
-            height: 660,
-        },
-        "35": {
-            advance: 1726,
-            bearingX: 32,
-            bearingY: 1576,
-            width: 1674,
-            height: 1604,
-        },
-        "36": {
-            advance: 1420,
-            bearingX: 155,
-            bearingY: 1721,
-            width: 1126,
-            height: 2117,
-        },
-        "37": {
-            advance: 1680,
-            bearingX: 147,
-            bearingY: 1642,
-            width: 1462,
-            height: 1672,
-        },
-        "38": {
-            advance: 1340,
-            bearingX: 75,
-            bearingY: 1564,
-            width: 1193,
-            height: 1658,
-        },
-        "39": {
-            advance: 795,
-            bearingX: 284,
-            bearingY: 1660,
-            width: 188,
-            height: 520,
-        },
-        "40": {
-            advance: 750,
-            bearingX: 114,
-            bearingY: 1606,
-            width: 582,
-            height: 2038,
-        },
-        "41": {
-            advance: 750,
-            bearingX: 114,
-            bearingY: 1606,
-            width: 582,
-            height: 2038,
-        },
-        "42": {
-            advance: 1085,
-            bearingX: 47,
-            bearingY: 1594,
-            width: 916,
-            height: 813,
-        },
-        "43": {
-            advance: 984,
-            bearingX: 48,
-            bearingY: 1046,
-            width: 865,
-            height: 814,
-        },
-        "44": {
-            advance: 567,
-            bearingX: 198,
-            bearingY: 142,
-            width: 301,
-            height: 486,
-        },
-        "45": {
-            advance: 853,
-            bearingX: 111,
-            bearingY: 631,
-            width: 647,
-            height: 170,
-        },
-        "46": {
-            advance: 510,
-            bearingX: 143,
-            bearingY: 155,
-            width: 248,
-            height: 248,
-        },
-        "47": {
-            advance: 1048,
-            bearingX: 84,
-            bearingY: 1629,
-            width: 887,
-            height: 1719,
-        },
-        "48": {
-            advance: 1250,
-            bearingX: 61,
-            bearingY: 1555,
-            width: 1118,
-            height: 1594,
-        },
-        "49": {
-            advance: 922,
-            bearingX: 158,
-            bearingY: 1559,
-            width: 643,
-            height: 1560,
-        },
-        "50": {
-            advance: 1250,
-            bearingX: 164,
-            bearingY: 1536,
-            width: 938,
-            height: 1539,
-        },
-        "51": {
-            advance: 1250,
-            bearingX: 146,
-            bearingY: 1526,
-            width: 937,
-            height: 1572,
-        },
-        "52": {
-            advance: 1250,
-            bearingX: 48,
-            bearingY: 1560,
-            width: 1141,
-            height: 1586,
-        },
-        "53": {
-            advance: 1250,
-            bearingX: 126,
-            bearingY: 1544,
-            width: 1027,
-            height: 1607,
-        },
-        "54": {
-            advance: 1250,
-            bearingX: 111,
-            bearingY: 1556,
-            width: 998,
-            height: 1628,
-        },
-        "55": {
-            advance: 1250,
-            bearingX: 70,
-            bearingY: 1508,
-            width: 1144,
-            height: 1574,
-        },
-        "56": {
-            advance: 1250,
-            bearingX: 118,
-            bearingY: 1530,
-            width: 1008,
-            height: 1584,
-        },
-        "57": {
-            advance: 1250,
-            bearingX: 112,
-            bearingY: 1536,
-            width: 1044,
-            height: 1631,
-        },
-        "58": {
-            advance: 612,
-            bearingX: 183,
-            bearingY: 1129,
-            width: 231,
-            height: 1006,
-        },
-        "59": {
-            advance: 612,
-            bearingX: 81,
-            bearingY: 1124,
-            width: 330,
-            height: 1318,
-        },
-        "60": {
-            advance: 781,
-            bearingX: 19,
-            bearingY: 1061,
-            width: 608,
-            height: 867,
-        },
-        "61": {
-            advance: 1045,
-            bearingX: 99,
-            bearingY: 999,
-            width: 781,
-            height: 723,
-        },
-        "62": {
-            advance: 781,
-            bearingX: 58,
-            bearingY: 1125,
-            width: 678,
-            height: 939,
-        },
-        "63": {
-            advance: 1073,
-            bearingX: 52,
-            bearingY: 1478,
-            width: 909,
-            height: 1550,
-        },
-        "64": {
-            advance: 1907,
-            bearingX: 110,
-            bearingY: 1630,
-            width: 1643,
-            height: 1770,
-        },
-        "65": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 1478,
-            width: 1221,
-            height: 1508,
-        },
-        "66": {
-            advance: 1291,
-            bearingX: 191,
-            bearingY: 1571,
-            width: 1017,
-            height: 1619,
-        },
-        "67": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 1523,
-            width: 1113,
-            height: 1547,
-        },
-        "68": {
-            advance: 1478,
-            bearingX: 183,
-            bearingY: 1556,
-            width: 1193,
-            height: 1655,
-        },
-        "69": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 1605,
-            width: 1073,
-            height: 1704,
-        },
-        "70": {
-            advance: 1243,
-            bearingX: 173,
-            bearingY: 1573,
-            width: 1031,
-            height: 1679,
-        },
-        "71": {
-            advance: 1392,
-            bearingX: 79,
-            bearingY: 1571,
-            width: 1276,
-            height: 1640,
-        },
-        "72": {
-            advance: 1573,
-            bearingX: 152,
-            bearingY: 1554,
-            width: 1313,
-            height: 1637,
-        },
-        "73": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1494,
-            width: 985,
-            height: 1530,
-        },
-        "74": {
-            advance: 1362,
-            bearingX: 94,
-            bearingY: 1515,
-            width: 1202,
-            height: 1635,
-        },
-        "75": {
-            advance: 1251,
-            bearingX: 214,
-            bearingY: 1531,
-            width: 1029,
-            height: 1641,
-        },
-        "76": {
-            advance: 1128,
-            bearingX: 101,
-            bearingY: 1542,
-            width: 988,
-            height: 1635,
-        },
-        "77": {
-            advance: 1808,
-            bearingX: 112,
-            bearingY: 1534,
-            width: 1619,
-            height: 1617,
-        },
-        "78": {
-            advance: 1632,
-            bearingX: 123,
-            bearingY: 1551,
-            width: 1424,
-            height: 1630,
-        },
-        "79": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 1515,
-            width: 1432,
-            height: 1576,
-        },
-        "80": {
-            advance: 1066,
-            bearingX: 100,
-            bearingY: 1571,
-            width: 905,
-            height: 1595,
-        },
-        "81": {
-            advance: 1795,
-            bearingX: 77,
-            bearingY: 1515,
-            width: 1673,
-            height: 1953,
-        },
-        "82": {
-            advance: 1287,
-            bearingX: 117,
-            bearingY: 1541,
-            width: 1111,
-            height: 1576,
-        },
-        "83": {
-            advance: 1420,
-            bearingX: 133,
-            bearingY: 1467,
-            width: 1188,
-            height: 1524,
-        },
-        "84": {
-            advance: 1392,
-            bearingX: 116,
-            bearingY: 1518,
-            width: 1347,
-            height: 1526,
-        },
-        "85": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 1503,
-            width: 1234,
-            height: 1543,
-        },
-        "86": {
-            advance: 1331,
-            bearingX: 145,
-            bearingY: 1535,
-            width: 1181,
-            height: 1615,
-        },
-        "87": {
-            advance: 2129,
-            bearingX: 139,
-            bearingY: 1526,
-            width: 1957,
-            height: 1622,
-        },
-        "88": {
-            advance: 1482,
-            bearingX: 68,
-            bearingY: 1521,
-            width: 1338,
-            height: 1605,
-        },
-        "89": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 1521,
-            width: 1196,
-            height: 1591,
-        },
-        "90": {
-            advance: 1420,
-            bearingX: 68,
-            bearingY: 1509,
-            width: 1314,
-            height: 1560,
-        },
-        "91": {
-            advance: 771,
-            bearingX: 176,
-            bearingY: 1520,
-            width: 526,
-            height: 1937,
-        },
-        "92": {
-            advance: 1126,
-            bearingX: 177,
-            bearingY: 1524,
-            width: 838,
-            height: 1665,
-        },
-        "93": {
-            advance: 771,
-            bearingX: 176,
-            bearingY: 1520,
-            width: 526,
-            height: 1937,
-        },
-        "94": {
-            advance: 1190,
-            bearingX: 198,
-            bearingY: 1646,
-            width: 823,
-            height: 524,
-        },
-        "95": {
-            advance: 1284,
-            bearingX: -35,
-            bearingY: -157,
-            width: 1358,
-            height: 188,
-        },
-        "96": {
-            advance: 1139,
-            bearingX: 149,
-            bearingY: 1661,
-            width: 428,
-            height: 483,
-        },
-        "97": {
-            advance: 1048,
-            bearingX: 51,
-            bearingY: 1044,
-            width: 962,
-            height: 1110,
-        },
-        "98": {
-            advance: 1215,
-            bearingX: 153,
-            bearingY: 1575,
-            width: 966,
-            height: 1617,
-        },
-        "99": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1063,
-            width: 864,
-            height: 1126,
-        },
-        "100": {
-            advance: 1203,
-            bearingX: 103,
-            bearingY: 1594,
-            width: 997,
-            height: 1640,
-        },
-        "101": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1045,
-            width: 996,
-            height: 1091,
-        },
-        "102": {
-            advance: 1041,
-            bearingX: 75,
-            bearingY: 1599,
-            width: 866,
-            height: 1759,
-        },
-        "103": {
-            advance: 1087,
-            bearingX: 58,
-            bearingY: 1024,
-            width: 952,
-            height: 1588,
-        },
-        "104": {
-            advance: 1183,
-            bearingX: 145,
-            bearingY: 1602,
-            width: 933,
-            height: 1664,
-        },
-        "105": {
-            advance: 574,
-            bearingX: 179,
-            bearingY: 1498,
-            width: 268,
-            height: 1503,
-        },
-        "106": {
-            advance: 826,
-            bearingX: -18,
-            bearingY: 1496,
-            width: 674,
-            height: 2093,
-        },
-        "107": {
-            advance: 1106,
-            bearingX: 163,
-            bearingY: 1604,
-            width: 913,
-            height: 1645,
-        },
-        "108": {
-            advance: 561,
-            bearingX: 174,
-            bearingY: 1609,
-            width: 229,
-            height: 1652,
-        },
-        "109": {
-            advance: 1591,
-            bearingX: 121,
-            bearingY: 1110,
-            width: 1391,
-            height: 1234,
-        },
-        "110": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1092,
-            width: 882,
-            height: 1163,
-        },
-        "111": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1037,
-            width: 895,
-            height: 1096,
-        },
-        "112": {
-            advance: 1095,
-            bearingX: 119,
-            bearingY: 1099,
-            width: 889,
-            height: 1680,
-        },
-        "113": {
-            advance: 1065,
-            bearingX: 60,
-            bearingY: 1065,
-            width: 883,
-            height: 1621,
-        },
-        "114": {
-            advance: 984,
-            bearingX: 139,
-            bearingY: 1054,
-            width: 779,
-            height: 1121,
-        },
-        "115": {
-            advance: 997,
-            bearingX: 41,
-            bearingY: 1142,
-            width: 871,
-            height: 1202,
-        },
-        "116": {
-            advance: 965,
-            bearingX: 65,
-            bearingY: 1397,
-            width: 841,
-            height: 1461,
-        },
-        "117": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1066,
-            width: 866,
-            height: 1146,
-        },
-        "118": {
-            advance: 996,
-            bearingX: 62,
-            bearingY: 1056,
-            width: 908,
-            height: 1096,
-        },
-        "119": {
-            advance: 1401,
-            bearingX: 76,
-            bearingY: 1041,
-            width: 1270,
-            height: 1123,
-        },
-        "120": {
-            advance: 1209,
-            bearingX: 61,
-            bearingY: 1105,
-            width: 1090,
-            height: 1149,
-        },
-        "121": {
-            advance: 1066,
-            bearingX: -4,
-            bearingY: 1040,
-            width: 1028,
-            height: 1619,
-        },
-        "122": {
-            advance: 1102,
-            bearingX: 122,
-            bearingY: 1056,
-            width: 918,
-            height: 1133,
-        },
-        "123": {
-            advance: 750,
-            bearingX: 6,
-            bearingY: 1625,
-            width: 691,
-            height: 2009,
-        },
-        "124": {
-            advance: 863,
-            bearingX: 353,
-            bearingY: 1715,
-            width: 178,
-            height: 2077,
-        },
-        "125": {
-            advance: 750,
-            bearingX: 6,
-            bearingY: 1625,
-            width: 691,
-            height: 2009,
-        },
-        "126": {
-            advance: 1224,
-            bearingX: 100,
-            bearingY: 934,
-            width: 1043,
-            height: 467,
-        },
-        "160": {
-            advance: 612,
-            bearingX: 0,
-            bearingY: 0,
-            width: 0,
-            height: 0,
-        },
-        "161": {
-            advance: 487,
-            bearingX: 120,
-            bearingY: 1605,
-            width: 212,
-            height: 1671,
-        },
-        "162": {
-            advance: 1276,
-            bearingX: 193,
-            bearingY: 1740,
-            width: 989,
-            height: 1708,
-        },
-        "163": {
-            advance: 1625,
-            bearingX: 37,
-            bearingY: 1610,
-            width: 1422,
-            height: 1760,
-        },
-        "164": {
-            advance: 1253,
-            bearingX: 56,
-            bearingY: 1319,
-            width: 1175,
-            height: 1139,
-        },
-        "165": {
-            advance: 1302,
-            bearingX: 133,
-            bearingY: 1458,
-            width: 1039,
-            height: 1511,
-        },
-        "166": {
-            advance: 828,
-            bearingX: 327,
-            bearingY: 1695,
-            width: 164,
-            height: 1739,
-        },
-        "167": {
-            advance: 1300,
-            bearingX: 117,
-            bearingY: 1641,
-            width: 983,
-            height: 1718,
-        },
-        "168": {
-            advance: 1139,
-            bearingX: 221,
-            bearingY: 1380,
-            width: 754,
-            height: 196,
-        },
-        "169": {
-            advance: 1629,
-            bearingX: 90,
-            bearingY: 1620,
-            width: 1432,
-            height: 1377,
-        },
-        "170": {
-            advance: 1079,
-            bearingX: 101,
-            bearingY: 1683,
-            width: 811,
-            height: 792,
-        },
-        "171": {
-            advance: 1183,
-            bearingX: 19,
-            bearingY: 1061,
-            width: 1062,
-            height: 889,
-        },
-        "172": {
-            advance: 984,
-            bearingX: 48,
-            bearingY: 737,
-            width: 867,
-            height: 489,
-        },
-        "173": {
-            advance: 853,
-            bearingX: 111,
-            bearingY: 631,
-            width: 647,
-            height: 170,
-        },
-        "174": {
-            advance: 1629,
-            bearingX: 91,
-            bearingY: 1631,
-            width: 1448,
-            height: 1424,
-        },
-        "175": {
-            advance: 1284,
-            bearingX: -35,
-            bearingY: 1891,
-            width: 1358,
-            height: 188,
-        },
-        "176": {
-            advance: 838,
-            bearingX: 68,
-            bearingY: 1680,
-            width: 698,
-            height: 706,
-        },
-        "177": {
-            advance: 984,
-            bearingX: 48,
-            bearingY: 1046,
-            width: 865,
-            height: 1174,
-        },
-        "178": {
-            advance: 1333,
-            bearingX: 381,
-            bearingY: 1748,
-            width: 603,
-            height: 845,
-        },
-        "179": {
-            advance: 1333,
-            bearingX: 414,
-            bearingY: 1701,
-            width: 533,
-            height: 846,
-        },
-        "180": {
-            advance: 1139,
-            bearingX: 213,
-            bearingY: 1665,
-            width: 452,
-            height: 478,
-        },
-        "181": {
-            advance: 1065,
-            bearingX: 22,
-            bearingY: 1066,
-            width: 1159,
-            height: 1472,
-        },
-        "182": {
-            advance: 1421,
-            bearingX: 71,
-            bearingY: 1657,
-            width: 1255,
-            height: 1853,
-        },
-        "183": {
-            advance: 510,
-            bearingX: 143,
-            bearingY: 840,
-            width: 248,
-            height: 248,
-        },
-        "184": {
-            advance: 1139,
-            bearingX: 366,
-            bearingY: 115,
-            width: 499,
-            height: 560,
-        },
-        "185": {
-            advance: 1333,
-            bearingX: 402,
-            bearingY: 1724,
-            width: 556,
-            height: 821,
-        },
-        "186": {
-            advance: 920,
-            bearingX: 40,
-            bearingY: 1702,
-            width: 821,
-            height: 810,
-        },
-        "187": {
-            advance: 1183,
-            bearingX: 19,
-            bearingY: 1061,
-            width: 1062,
-            height: 889,
-        },
-        "188": {
-            advance: 1333,
-            bearingX: 162,
-            bearingY: 1724,
-            width: 1049,
-            height: 2066,
-        },
-        "189": {
-            advance: 1333,
-            bearingX: 162,
-            bearingY: 1724,
-            width: 1049,
-            height: 2082,
-        },
-        "190": {
-            advance: 1333,
-            bearingX: 162,
-            bearingY: 1701,
-            width: 1049,
-            height: 2043,
-        },
-        "191": {
-            advance: 1073,
-            bearingX: 52,
-            bearingY: 1478,
-            width: 909,
-            height: 1550,
-        },
-        "192": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 2121,
-            width: 1221,
-            height: 2151,
-        },
-        "193": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 2108,
-            width: 1221,
-            height: 2138,
-        },
-        "194": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 2090,
-            width: 1221,
-            height: 2120,
-        },
-        "195": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 2005,
-            width: 1316,
-            height: 2035,
-        },
-        "196": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 1830,
-            width: 1221,
-            height: 1860,
-        },
-        "197": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 2001,
-            width: 1221,
-            height: 2031,
-        },
-        "198": {
-            advance: 2226,
-            bearingX: 46,
-            bearingY: 1621,
-            width: 2170,
-            height: 1708,
-        },
-        "199": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 1523,
-            width: 1113,
-            height: 1930,
-        },
-        "200": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 2230,
-            width: 1073,
-            height: 2329,
-        },
-        "201": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 2242,
-            width: 1073,
-            height: 2341,
-        },
-        "202": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 2257,
-            width: 1073,
-            height: 2356,
-        },
-        "203": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 1948,
-            width: 1073,
-            height: 2047,
-        },
-        "204": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 2161,
-            width: 985,
-            height: 2197,
-        },
-        "205": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 2135,
-            width: 985,
-            height: 2171,
-        },
-        "206": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 2177,
-            width: 985,
-            height: 2213,
-        },
-        "207": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1850,
-            width: 985,
-            height: 1886,
-        },
-        "208": {
-            advance: 1478,
-            bearingX: -44,
-            bearingY: 1556,
-            width: 1420,
-            height: 1655,
-        },
-        "209": {
-            advance: 1632,
-            bearingX: 123,
-            bearingY: 2024,
-            width: 1424,
-            height: 2103,
-        },
-        "210": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 2131,
-            width: 1432,
-            height: 2192,
-        },
-        "211": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 2115,
-            width: 1432,
-            height: 2176,
-        },
-        "212": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 2140,
-            width: 1432,
-            height: 2201,
-        },
-        "213": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 2007,
-            width: 1432,
-            height: 2068,
-        },
-        "214": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 1856,
-            width: 1432,
-            height: 1917,
-        },
-        "215": {
-            advance: 984,
-            bearingX: 67,
-            bearingY: 1020,
-            width: 784,
-            height: 796,
-        },
-        "216": {
-            advance: 1635,
-            bearingX: 83,
-            bearingY: 1542,
-            width: 1548,
-            height: 1632,
-        },
-        "217": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 2117,
-            width: 1234,
-            height: 2157,
-        },
-        "218": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 2121,
-            width: 1234,
-            height: 2161,
-        },
-        "219": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 2146,
-            width: 1234,
-            height: 2186,
-        },
-        "220": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 1833,
-            width: 1234,
-            height: 1873,
-        },
-        "221": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 2206,
-            width: 1196,
-            height: 2276,
-        },
-        "222": {
-            advance: 1066,
-            bearingX: 100,
-            bearingY: 1552,
-            width: 911,
-            height: 1576,
-        },
-        "223": {
-            advance: 908,
-            bearingX: 46,
-            bearingY: 1570,
-            width: 839,
-            height: 1727,
-        },
-        "224": {
-            advance: 1048,
-            bearingX: 66,
-            bearingY: 1654,
-            width: 962,
-            height: 1720,
-        },
-        "225": {
-            advance: 1048,
-            bearingX: 66,
-            bearingY: 1665,
-            width: 962,
-            height: 1731,
-        },
-        "226": {
-            advance: 1048,
-            bearingX: 66,
-            bearingY: 1698,
-            width: 962,
-            height: 1764,
-        },
-        "227": {
-            advance: 1048,
-            bearingX: 66,
-            bearingY: 1534,
-            width: 1001,
-            height: 1600,
-        },
-        "228": {
-            advance: 1048,
-            bearingX: 66,
-            bearingY: 1380,
-            width: 962,
-            height: 1446,
-        },
-        "229": {
-            advance: 1048,
-            bearingX: 66,
-            bearingY: 1758,
-            width: 962,
-            height: 1824,
-        },
-        "230": {
-            advance: 1867,
-            bearingX: 67,
-            bearingY: 1045,
-            width: 1726,
-            height: 1094,
-        },
-        "231": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1063,
-            width: 864,
-            height: 1508,
-        },
-        "232": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1661,
-            width: 996,
-            height: 1707,
-        },
-        "233": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1665,
-            width: 996,
-            height: 1711,
-        },
-        "234": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1690,
-            width: 996,
-            height: 1736,
-        },
-        "235": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1477,
-            width: 996,
-            height: 1523,
-        },
-        "236": {
-            advance: 574,
-            bearingX: 25,
-            bearingY: 1661,
-            width: 428,
-            height: 1666,
-        },
-        "237": {
-            advance: 574,
-            bearingX: 96,
-            bearingY: 1665,
-            width: 452,
-            height: 1670,
-        },
-        "238": {
-            advance: 574,
-            bearingX: -142,
-            bearingY: 1640,
-            width: 822,
-            height: 1645,
-        },
-        "239": {
-            advance: 574,
-            bearingX: -122,
-            bearingY: 1380,
-            width: 754,
-            height: 1385,
-        },
-        "240": {
-            advance: 1042,
-            bearingX: 185,
-            bearingY: 1538,
-            width: 806,
-            height: 1579,
-        },
-        "241": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1554,
-            width: 950,
-            height: 1625,
-        },
-        "242": {
-            advance: 1077,
-            bearingX: 83,
-            bearingY: 1661,
-            width: 895,
-            height: 1720,
-        },
-        "243": {
-            advance: 1077,
-            bearingX: 83,
-            bearingY: 1665,
-            width: 895,
-            height: 1724,
-        },
-        "244": {
-            advance: 1077,
-            bearingX: 83,
-            bearingY: 1690,
-            width: 911,
-            height: 1749,
-        },
-        "245": {
-            advance: 1077,
-            bearingX: 83,
-            bearingY: 1554,
-            width: 1010,
-            height: 1613,
-        },
-        "246": {
-            advance: 1077,
-            bearingX: 83,
-            bearingY: 1380,
-            width: 936,
-            height: 1439,
-        },
-        "247": {
-            advance: 984,
-            bearingX: 48,
-            bearingY: 1140,
-            width: 865,
-            height: 953,
-        },
-        "248": {
-            advance: 1077,
-            bearingX: 53,
-            bearingY: 1046,
-            width: 977,
-            height: 1105,
-        },
-        "249": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1669,
-            width: 866,
-            height: 1749,
-        },
-        "250": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1665,
-            width: 866,
-            height: 1745,
-        },
-        "251": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1690,
-            width: 866,
-            height: 1770,
-        },
-        "252": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1472,
-            width: 866,
-            height: 1552,
-        },
-        "253": {
-            advance: 1066,
-            bearingX: -4,
-            bearingY: 1665,
-            width: 1028,
-            height: 2244,
-        },
-        "254": {
-            advance: 1095,
-            bearingX: 112,
-            bearingY: 1494,
-            width: 896,
-            height: 2075,
-        },
-        "255": {
-            advance: 883,
-            bearingX: -4,
-            bearingY: 1380,
-            width: 1028,
-            height: 1959,
-        },
-        "256": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 1794,
-            width: 1221,
-            height: 1824,
-        },
-        "257": {
-            advance: 1048,
-            bearingX: 51,
-            bearingY: 1344,
-            width: 962,
-            height: 1410,
-        },
-        "258": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 2039,
-            width: 1228,
-            height: 2069,
-        },
-        "259": {
-            advance: 1048,
-            bearingX: 51,
-            bearingY: 1589,
-            width: 962,
-            height: 1655,
-        },
-        "260": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 1478,
-            width: 1440,
-            height: 1822,
-        },
-        "261": {
-            advance: 1048,
-            bearingX: 51,
-            bearingY: 1044,
-            width: 1259,
-            height: 1388,
-        },
-        "262": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 2115,
-            width: 1113,
-            height: 2139,
-        },
-        "263": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1665,
-            width: 864,
-            height: 1728,
-        },
-        "264": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 2140,
-            width: 1116,
-            height: 2164,
-        },
-        "265": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1590,
-            width: 911,
-            height: 1653,
-        },
-        "266": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 1854,
-            width: 1113,
-            height: 1878,
-        },
-        "267": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1404,
-            width: 864,
-            height: 1467,
-        },
-        "268": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 2077,
-            width: 1113,
-            height: 2101,
-        },
-        "269": {
-            advance: 1234,
-            bearingX: 105,
-            bearingY: 1627,
-            width: 864,
-            height: 1690,
-        },
-        "270": {
-            advance: 1478,
-            bearingX: 183,
-            bearingY: 2077,
-            width: 1193,
-            height: 2176,
-        },
-        "271": {
-            advance: 1698,
-            bearingX: 103,
-            bearingY: 1603,
-            width: 1404,
-            height: 1649,
-        },
-        "272": {
-            advance: 1478,
-            bearingX: -44,
-            bearingY: 1556,
-            width: 1420,
-            height: 1655,
-        },
-        "273": {
-            advance: 1234,
-            bearingX: 103,
-            bearingY: 1594,
-            width: 1118,
-            height: 1640,
-        },
-        "274": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 1874,
-            width: 1073,
-            height: 1973,
-        },
-        "275": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1344,
-            width: 996,
-            height: 1390,
-        },
-        "276": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 2089,
-            width: 1073,
-            height: 2188,
-        },
-        "277": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1589,
-            width: 996,
-            height: 1635,
-        },
-        "278": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 1904,
-            width: 1073,
-            height: 2003,
-        },
-        "279": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1404,
-            width: 996,
-            height: 1450,
-        },
-        "280": {
-            advance: 1279,
-            bearingX: 200,
-            bearingY: 1605,
-            width: 1073,
-            height: 2019,
-        },
-        "281": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1045,
-            width: 996,
-            height: 1429,
-        },
-        "282": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 2107,
-            width: 1073,
-            height: 2206,
-        },
-        "283": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1627,
-            width: 996,
-            height: 1673,
-        },
-        "284": {
-            advance: 1392,
-            bearingX: 79,
-            bearingY: 2140,
-            width: 1276,
-            height: 2209,
-        },
-        "285": {
-            advance: 1087,
-            bearingX: 58,
-            bearingY: 1690,
-            width: 978,
-            height: 2254,
-        },
-        "286": {
-            advance: 1392,
-            bearingX: 79,
-            bearingY: 2039,
-            width: 1276,
-            height: 2108,
-        },
-        "287": {
-            advance: 1087,
-            bearingX: 58,
-            bearingY: 1589,
-            width: 1001,
-            height: 2153,
-        },
-        "288": {
-            advance: 1392,
-            bearingX: 79,
-            bearingY: 1854,
-            width: 1276,
-            height: 1923,
-        },
-        "289": {
-            advance: 1087,
-            bearingX: 58,
-            bearingY: 1404,
-            width: 952,
-            height: 1968,
-        },
-        "290": {
-            advance: 1392,
-            bearingX: 79,
-            bearingY: 1571,
-            width: 1276,
-            height: 2016,
-        },
-        "291": {
-            advance: 1087,
-            bearingX: 58,
-            bearingY: 1693,
-            width: 952,
-            height: 2257,
-        },
-        "292": {
-            advance: 1573,
-            bearingX: 152,
-            bearingY: 2140,
-            width: 1313,
-            height: 2223,
-        },
-        "293": {
-            advance: 1183,
-            bearingX: 145,
-            bearingY: 2140,
-            width: 991,
-            height: 2202,
-        },
-        "294": {
-            advance: 1573,
-            bearingX: -35,
-            bearingY: 1554,
-            width: 1642,
-            height: 1637,
-        },
-        "295": {
-            advance: 1183,
-            bearingX: -12,
-            bearingY: 1602,
-            width: 1090,
-            height: 1664,
-        },
-        "296": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 2004,
-            width: 985,
-            height: 2040,
-        },
-        "297": {
-            advance: 574,
-            bearingX: -190,
-            bearingY: 1554,
-            width: 926,
-            height: 1559,
-        },
-        "298": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1794,
-            width: 985,
-            height: 1830,
-        },
-        "299": {
-            advance: 574,
-            bearingX: -82,
-            bearingY: 1344,
-            width: 693,
-            height: 1349,
-        },
-        "300": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 2039,
-            width: 985,
-            height: 2075,
-        },
-        "301": {
-            advance: 574,
-            bearingX: -127,
-            bearingY: 1589,
-            width: 846,
-            height: 1594,
-        },
-        "302": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1494,
-            width: 985,
-            height: 1838,
-        },
-        "303": {
-            advance: 574,
-            bearingX: 48,
-            bearingY: 1498,
-            width: 523,
-            height: 1842,
-        },
-        "304": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1854,
-            width: 985,
-            height: 1890,
-        },
-        "305": {
-            advance: 574,
-            bearingX: 179,
-            bearingY: 1024,
-            width: 216,
-            height: 1029,
-        },
-        "306": {
-            advance: 2307,
-            bearingX: 75,
-            bearingY: 1515,
-            width: 2165,
-            height: 1635,
-        },
-        "307": {
-            advance: 1086,
-            bearingX: 179,
-            bearingY: 1498,
-            width: 737,
-            height: 2095,
-        },
-        "308": {
-            advance: 1362,
-            bearingX: 94,
-            bearingY: 2140,
-            width: 1202,
-            height: 2260,
-        },
-        "309": {
-            advance: 826,
-            bearingX: -56,
-            bearingY: 1690,
-            width: 822,
-            height: 2287,
-        },
-        "310": {
-            advance: 1251,
-            bearingX: 214,
-            bearingY: 1531,
-            width: 1029,
-            height: 1976,
-        },
-        "311": {
-            advance: 1106,
-            bearingX: 163,
-            bearingY: 1604,
-            width: 913,
-            height: 2049,
-        },
-        "312": {
-            advance: 1106,
-            bearingX: 125,
-            bearingY: 1164,
-            width: 913,
-            height: 1205,
-        },
-        "313": {
-            advance: 1128,
-            bearingX: 101,
-            bearingY: 2115,
-            width: 988,
-            height: 2208,
-        },
-        "314": {
-            advance: 561,
-            bearingX: 174,
-            bearingY: 2165,
-            width: 491,
-            height: 2208,
-        },
-        "315": {
-            advance: 1128,
-            bearingX: 101,
-            bearingY: 1542,
-            width: 988,
-            height: 1987,
-        },
-        "316": {
-            advance: 561,
-            bearingX: 174,
-            bearingY: 1609,
-            width: 501,
-            height: 2054,
-        },
-        "317": {
-            advance: 1128,
-            bearingX: 101,
-            bearingY: 1542,
-            width: 988,
-            height: 1635,
-        },
-        "318": {
-            advance: 951,
-            bearingX: 174,
-            bearingY: 1609,
-            width: 619,
-            height: 1652,
-        },
-        "319": {
-            advance: 1128,
-            bearingX: 101,
-            bearingY: 1542,
-            width: 988,
-            height: 1635,
-        },
-        "320": {
-            advance: 809,
-            bearingX: 174,
-            bearingY: 1609,
-            width: 637,
-            height: 1652,
-        },
-        "321": {
-            advance: 1128,
-            bearingX: -129,
-            bearingY: 1542,
-            width: 1218,
-            height: 1635,
-        },
-        "322": {
-            advance: 465,
-            bearingX: -87,
-            bearingY: 1519,
-            width: 648,
-            height: 1554,
-        },
-        "323": {
-            advance: 1632,
-            bearingX: 123,
-            bearingY: 2115,
-            width: 1424,
-            height: 2194,
-        },
-        "324": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1665,
-            width: 882,
-            height: 1736,
-        },
-        "325": {
-            advance: 1632,
-            bearingX: 123,
-            bearingY: 1551,
-            width: 1424,
-            height: 1996,
-        },
-        "326": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1092,
-            width: 882,
-            height: 1537,
-        },
-        "327": {
-            advance: 1632,
-            bearingX: 123,
-            bearingY: 2077,
-            width: 1424,
-            height: 2156,
-        },
-        "328": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1627,
-            width: 882,
-            height: 1698,
-        },
-        "329": {
-            advance: 1272,
-            bearingX: 57,
-            bearingY: 1584,
-            width: 1149,
-            height: 1655,
-        },
-        "330": {
-            advance: 1632,
-            bearingX: 123,
-            bearingY: 1551,
-            width: 1424,
-            height: 2077,
-        },
-        "331": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1092,
-            width: 892,
-            height: 1689,
-        },
-        "332": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 1794,
-            width: 1432,
-            height: 1855,
-        },
-        "333": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1344,
-            width: 895,
-            height: 1403,
-        },
-        "334": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 2039,
-            width: 1432,
-            height: 2100,
-        },
-        "335": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1589,
-            width: 962,
-            height: 1648,
-        },
-        "336": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 2045,
-            width: 1432,
-            height: 2106,
-        },
-        "337": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1595,
-            width: 922,
-            height: 1654,
-        },
-        "338": {
-            advance: 2445,
-            bearingX: 130,
-            bearingY: 1628,
-            width: 2300,
-            height: 1704,
-        },
-        "339": {
-            advance: 1837,
-            bearingX: 77,
-            bearingY: 1045,
-            width: 1721,
-            height: 1104,
-        },
-        "340": {
-            advance: 1287,
-            bearingX: 117,
-            bearingY: 2165,
-            width: 1111,
-            height: 2200,
-        },
-        "341": {
-            advance: 984,
-            bearingX: 139,
-            bearingY: 1665,
-            width: 779,
-            height: 1732,
-        },
-        "342": {
-            advance: 1287,
-            bearingX: 117,
-            bearingY: 1541,
-            width: 1111,
-            height: 1986,
-        },
-        "343": {
-            advance: 984,
-            bearingX: 59,
-            bearingY: 1054,
-            width: 859,
-            height: 1499,
-        },
-        "344": {
-            advance: 1287,
-            bearingX: 117,
-            bearingY: 2077,
-            width: 1111,
-            height: 2112,
-        },
-        "345": {
-            advance: 984,
-            bearingX: 139,
-            bearingY: 1627,
-            width: 779,
-            height: 1694,
-        },
-        "346": {
-            advance: 1420,
-            bearingX: 133,
-            bearingY: 2115,
-            width: 1188,
-            height: 2172,
-        },
-        "347": {
-            advance: 997,
-            bearingX: 41,
-            bearingY: 1665,
-            width: 871,
-            height: 1725,
-        },
-        "348": {
-            advance: 1420,
-            bearingX: 133,
-            bearingY: 2140,
-            width: 1188,
-            height: 2197,
-        },
-        "349": {
-            advance: 997,
-            bearingX: 41,
-            bearingY: 1690,
-            width: 871,
-            height: 1750,
-        },
-        "350": {
-            advance: 1420,
-            bearingX: 133,
-            bearingY: 1467,
-            width: 1188,
-            height: 1912,
-        },
-        "351": {
-            advance: 997,
-            bearingX: 41,
-            bearingY: 1142,
-            width: 871,
-            height: 1587,
-        },
-        "352": {
-            advance: 1420,
-            bearingX: 134,
-            bearingY: 2078,
-            width: 1188,
-            height: 2135,
-        },
-        "353": {
-            advance: 826,
-            bearingX: 43,
-            bearingY: 1627,
-            width: 871,
-            height: 1687,
-        },
-        "354": {
-            advance: 1392,
-            bearingX: 116,
-            bearingY: 1518,
-            width: 1347,
-            height: 1886,
-        },
-        "355": {
-            advance: 965,
-            bearingX: 65,
-            bearingY: 1397,
-            width: 841,
-            height: 1811,
-        },
-        "356": {
-            advance: 1392,
-            bearingX: 116,
-            bearingY: 2127,
-            width: 1347,
-            height: 2135,
-        },
-        "357": {
-            advance: 1309,
-            bearingX: 65,
-            bearingY: 1397,
-            width: 1178,
-            height: 1461,
-        },
-        "358": {
-            advance: 1392,
-            bearingX: 42,
-            bearingY: 1515,
-            width: 1347,
-            height: 1523,
-        },
-        "359": {
-            advance: 965,
-            bearingX: 65,
-            bearingY: 1397,
-            width: 841,
-            height: 1461,
-        },
-        "360": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 2004,
-            width: 1234,
-            height: 2044,
-        },
-        "361": {
-            advance: 1065,
-            bearingX: 90,
-            bearingY: 1554,
-            width: 926,
-            height: 1634,
-        },
-        "362": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 1794,
-            width: 1234,
-            height: 1834,
-        },
-        "363": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1344,
-            width: 866,
-            height: 1424,
-        },
-        "364": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 2039,
-            width: 1234,
-            height: 2079,
-        },
-        "365": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1589,
-            width: 870,
-            height: 1669,
-        },
-        "366": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 2208,
-            width: 1234,
-            height: 2248,
-        },
-        "367": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1758,
-            width: 866,
-            height: 1838,
-        },
-        "368": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 2045,
-            width: 1234,
-            height: 2085,
-        },
-        "369": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1595,
-            width: 920,
-            height: 1675,
-        },
-        "370": {
-            advance: 1509,
-            bearingX: 155,
-            bearingY: 1503,
-            width: 1234,
-            height: 1947,
-        },
-        "371": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1066,
-            width: 1012,
-            height: 1510,
-        },
-        "372": {
-            advance: 2129,
-            bearingX: 139,
-            bearingY: 2140,
-            width: 1957,
-            height: 2236,
-        },
-        "373": {
-            advance: 1401,
-            bearingX: 76,
-            bearingY: 1690,
-            width: 1270,
-            height: 1772,
-        },
-        "374": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 2140,
-            width: 1196,
-            height: 2210,
-        },
-        "375": {
-            advance: 1066,
-            bearingX: -4,
-            bearingY: 1690,
-            width: 1028,
-            height: 2269,
-        },
-        "376": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 1835,
-            width: 1196,
-            height: 1905,
-        },
-        "377": {
-            advance: 1420,
-            bearingX: 68,
-            bearingY: 2165,
-            width: 1314,
-            height: 2216,
-        },
-        "378": {
-            advance: 1102,
-            bearingX: 122,
-            bearingY: 1665,
-            width: 918,
-            height: 1742,
-        },
-        "379": {
-            advance: 1420,
-            bearingX: 68,
-            bearingY: 1854,
-            width: 1314,
-            height: 1905,
-        },
-        "380": {
-            advance: 1102,
-            bearingX: 122,
-            bearingY: 1404,
-            width: 918,
-            height: 1481,
-        },
-        "381": {
-            advance: 1420,
-            bearingX: 68,
-            bearingY: 2175,
-            width: 1314,
-            height: 2226,
-        },
-        "382": {
-            advance: 1102,
-            bearingX: 122,
-            bearingY: 1627,
-            width: 918,
-            height: 1704,
-        },
-        "383": {
-            advance: 1041,
-            bearingX: 314,
-            bearingY: 1599,
-            width: 627,
-            height: 1759,
-        },
-        "402": {
-            advance: 874,
-            bearingX: -103,
-            bearingY: 1561,
-            width: 1250,
-            height: 1922,
-        },
-        "506": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 2255,
-            width: 1221,
-            height: 2285,
-        },
-        "507": {
-            advance: 1048,
-            bearingX: 66,
-            bearingY: 2005,
-            width: 962,
-            height: 2071,
-        },
-        "508": {
-            advance: 2226,
-            bearingX: 46,
-            bearingY: 2165,
-            width: 2170,
-            height: 2252,
-        },
-        "509": {
-            advance: 1867,
-            bearingX: 67,
-            bearingY: 1665,
-            width: 1726,
-            height: 1714,
-        },
-        "510": {
-            advance: 1635,
-            bearingX: 83,
-            bearingY: 2115,
-            width: 1548,
-            height: 2205,
-        },
-        "511": {
-            advance: 1077,
-            bearingX: 53,
-            bearingY: 1665,
-            width: 977,
-            height: 1724,
-        },
-        "536": {
-            advance: 1420,
-            bearingX: 133,
-            bearingY: 1467,
-            width: 1188,
-            height: 2106,
-        },
-        "537": {
-            advance: 997,
-            bearingX: 41,
-            bearingY: 1142,
-            width: 871,
-            height: 1781,
-        },
-        "538": {
-            advance: 1392,
-            bearingX: 116,
-            bearingY: 1518,
-            width: 1347,
-            height: 2157,
-        },
-        "539": {
-            advance: 965,
-            bearingX: 65,
-            bearingY: 1397,
-            width: 841,
-            height: 2036,
-        },
-        "710": {
-            advance: 1139,
-            bearingX: 194,
-            bearingY: 1690,
-            width: 822,
-            height: 524,
-        },
-        "711": {
-            advance: 1139,
-            bearingX: 284,
-            bearingY: 1627,
-            width: 670,
-            height: 451,
-        },
-        "713": {
-            advance: 1139,
-            bearingX: 278,
-            bearingY: 1344,
-            width: 693,
-            height: 158,
-        },
-        "728": {
-            advance: 1139,
-            bearingX: 213,
-            bearingY: 1589,
-            width: 846,
-            height: 410,
-        },
-        "729": {
-            advance: 1139,
-            bearingX: 316,
-            bearingY: 1404,
-            width: 244,
-            height: 221,
-        },
-        "730": {
-            advance: 1139,
-            bearingX: 263,
-            bearingY: 1758,
-            width: 674,
-            height: 574,
-        },
-        "731": {
-            advance: 1139,
-            bearingX: 298,
-            bearingY: 170,
-            width: 523,
-            height: 514,
-        },
-        "732": {
-            advance: 1139,
-            bearingX: 170,
-            bearingY: 1554,
-            width: 926,
-            height: 371,
-        },
-        "733": {
-            advance: 1139,
-            bearingX: 210,
-            bearingY: 1595,
-            width: 739,
-            height: 412,
-        },
-        "900": {
-            advance: 1139,
-            bearingX: 465,
-            bearingY: 1505,
-            width: 271,
-            height: 356,
-        },
-        "901": {
-            advance: 1139,
-            bearingX: 121,
-            bearingY: 1505,
-            width: 934,
-            height: 356,
-        },
-        "902": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 1555,
-            width: 1221,
-            height: 1585,
-        },
-        "903": {
-            advance: 510,
-            bearingX: 143,
-            bearingY: 840,
-            width: 248,
-            height: 248,
-        },
-        "904": {
-            advance: 1616,
-            bearingX: 142,
-            bearingY: 1605,
-            width: 1408,
-            height: 1704,
-        },
-        "905": {
-            advance: 2013,
-            bearingX: 195,
-            bearingY: 1555,
-            width: 1710,
-            height: 1638,
-        },
-        "906": {
-            advance: 1549,
-            bearingX: 44,
-            bearingY: 1555,
-            width: 1446,
-            height: 1591,
-        },
-        "908": {
-            advance: 1795,
-            bearingX: 95,
-            bearingY: 1555,
-            width: 1613,
-            height: 1616,
-        },
-        "910": {
-            advance: 1741,
-            bearingX: 45,
-            bearingY: 1555,
-            width: 1620,
-            height: 1625,
-        },
-        "911": {
-            advance: 1986,
-            bearingX: 65,
-            bearingY: 1630,
-            width: 1856,
-            height: 1770,
-        },
-        "912": {
-            advance: 574,
-            bearingX: -109,
-            bearingY: 1505,
-            width: 934,
-            height: 1510,
-        },
-        "913": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 1478,
-            width: 1221,
-            height: 1508,
-        },
-        "914": {
-            advance: 1291,
-            bearingX: 191,
-            bearingY: 1571,
-            width: 1017,
-            height: 1619,
-        },
-        "915": {
-            advance: 1262,
-            bearingX: 182,
-            bearingY: 1515,
-            width: 1007,
-            height: 1523,
-        },
-        "916": {
-            advance: 1627,
-            bearingX: 98,
-            bearingY: 1663,
-            width: 1503,
-            height: 1735,
-        },
-        "917": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 1605,
-            width: 1073,
-            height: 1704,
-        },
-        "918": {
-            advance: 1420,
-            bearingX: 68,
-            bearingY: 1509,
-            width: 1314,
-            height: 1560,
-        },
-        "919": {
-            advance: 1573,
-            bearingX: 152,
-            bearingY: 1554,
-            width: 1313,
-            height: 1637,
-        },
-        "920": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 1515,
-            width: 1432,
-            height: 1576,
-        },
-        "921": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1494,
-            width: 985,
-            height: 1530,
-        },
-        "922": {
-            advance: 1251,
-            bearingX: 214,
-            bearingY: 1531,
-            width: 1029,
-            height: 1641,
-        },
-        "923": {
-            advance: 1412,
-            bearingX: 16,
-            bearingY: 1556,
-            width: 1403,
-            height: 1636,
-        },
-        "924": {
-            advance: 1808,
-            bearingX: 112,
-            bearingY: 1534,
-            width: 1619,
-            height: 1617,
-        },
-        "925": {
-            advance: 1632,
-            bearingX: 123,
-            bearingY: 1551,
-            width: 1424,
-            height: 1630,
-        },
-        "926": {
-            advance: 1475,
-            bearingX: 86,
-            bearingY: 1506,
-            width: 1332,
-            height: 1538,
-        },
-        "927": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 1515,
-            width: 1432,
-            height: 1576,
-        },
-        "928": {
-            advance: 1860,
-            bearingX: 105,
-            bearingY: 1568,
-            width: 1651,
-            height: 1605,
-        },
-        "929": {
-            advance: 1066,
-            bearingX: 100,
-            bearingY: 1571,
-            width: 905,
-            height: 1595,
-        },
-        "931": {
-            advance: 1530,
-            bearingX: 123,
-            bearingY: 1521,
-            width: 1267,
-            height: 1548,
-        },
-        "932": {
-            advance: 1392,
-            bearingX: 116,
-            bearingY: 1518,
-            width: 1347,
-            height: 1526,
-        },
-        "933": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 1521,
-            width: 1196,
-            height: 1591,
-        },
-        "934": {
-            advance: 1307,
-            bearingX: 91,
-            bearingY: 1494,
-            width: 1181,
-            height: 1530,
-        },
-        "935": {
-            advance: 1482,
-            bearingX: 68,
-            bearingY: 1521,
-            width: 1338,
-            height: 1605,
-        },
-        "936": {
-            advance: 1380,
-            bearingX: 70,
-            bearingY: 1547,
-            width: 1261,
-            height: 1573,
-        },
-        "937": {
-            advance: 1966,
-            bearingX: 65,
-            bearingY: 1630,
-            width: 1836,
-            height: 1770,
-        },
-        "938": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1850,
-            width: 985,
-            height: 1886,
-        },
-        "939": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 1835,
-            width: 1196,
-            height: 1905,
-        },
-        "940": {
-            advance: 1091,
-            bearingX: 51,
-            bearingY: 1505,
-            width: 982,
-            height: 1571,
-        },
-        "941": {
-            advance: 1007,
-            bearingX: 156,
-            bearingY: 1505,
-            width: 754,
-            height: 1557,
-        },
-        "942": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1505,
-            width: 902,
-            height: 1826,
-        },
-        "943": {
-            advance: 574,
-            bearingX: 175,
-            bearingY: 1505,
-            width: 271,
-            height: 1510,
-        },
-        "944": {
-            advance: 1065,
-            bearingX: 74,
-            bearingY: 1505,
-            width: 981,
-            height: 1567,
-        },
-        "945": {
-            advance: 1091,
-            bearingX: 51,
-            bearingY: 1090,
-            width: 982,
-            height: 1156,
-        },
-        "946": {
-            advance: 1201,
-            bearingX: 111,
-            bearingY: 1480,
-            width: 984,
-            height: 1883,
-        },
-        "947": {
-            advance: 996,
-            bearingX: 62,
-            bearingY: 1056,
-            width: 908,
-            height: 1390,
-        },
-        "948": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1519,
-            width: 927,
-            height: 1578,
-        },
-        "949": {
-            advance: 1007,
-            bearingX: 156,
-            bearingY: 1052,
-            width: 754,
-            height: 1104,
-        },
-        "950": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1549,
-            width: 911,
-            height: 1964,
-        },
-        "951": {
-            advance: 1072,
-            bearingX: 124,
-            bearingY: 1092,
-            width: 902,
-            height: 1413,
-        },
-        "952": {
-            advance: 1250,
-            bearingX: 61,
-            bearingY: 1555,
-            width: 1118,
-            height: 1594,
-        },
-        "953": {
-            advance: 574,
-            bearingX: 179,
-            bearingY: 1024,
-            width: 216,
-            height: 1029,
-        },
-        "954": {
-            advance: 1106,
-            bearingX: 123,
-            bearingY: 1068,
-            width: 913,
-            height: 1109,
-        },
-        "955": {
-            advance: 996,
-            bearingX: 62,
-            bearingY: 1366,
-            width: 908,
-            height: 1390,
-        },
-        "956": {
-            advance: 1065,
-            bearingX: 109,
-            bearingY: 1066,
-            width: 866,
-            height: 1467,
-        },
-        "957": {
-            advance: 996,
-            bearingX: 62,
-            bearingY: 1056,
-            width: 908,
-            height: 1096,
-        },
-        "958": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1592,
-            width: 807,
-            height: 2007,
-        },
-        "959": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1037,
-            width: 895,
-            height: 1096,
-        },
-        "960": {
-            advance: 1286,
-            bearingX: 37,
-            bearingY: 1048,
-            width: 1217,
-            height: 1090,
-        },
-        "961": {
-            advance: 1077,
-            bearingX: 31,
-            bearingY: 1037,
-            width: 938,
-            height: 1451,
-        },
-        "962": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1063,
-            width: 845,
-            height: 1478,
-        },
-        "963": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1037,
-            width: 1040,
-            height: 1096,
-        },
-        "964": {
-            advance: 965,
-            bearingX: 65,
-            bearingY: 1057,
-            width: 841,
-            height: 1121,
-        },
-        "965": {
-            advance: 1065,
-            bearingX: 74,
-            bearingY: 1078,
-            width: 923,
-            height: 1140,
-        },
-        "966": {
-            advance: 1260,
-            bearingX: 114,
-            bearingY: 1031,
-            width: 1031,
-            height: 1431,
-        },
-        "967": {
-            advance: 1209,
-            bearingX: 32,
-            bearingY: 1073,
-            width: 1055,
-            height: 1445,
-        },
-        "968": {
-            advance: 1199,
-            bearingX: 64,
-            bearingY: 1091,
-            width: 1072,
-            height: 1500,
-        },
-        "969": {
-            advance: 1557,
-            bearingX: 56,
-            bearingY: 1068,
-            width: 1398,
-            height: 1110,
-        },
-        "970": {
-            advance: 574,
-            bearingX: -122,
-            bearingY: 1380,
-            width: 754,
-            height: 1385,
-        },
-        "971": {
-            advance: 1065,
-            bearingX: 74,
-            bearingY: 1380,
-            width: 923,
-            height: 1442,
-        },
-        "972": {
-            advance: 1077,
-            bearingX: 77,
-            bearingY: 1505,
-            width: 895,
-            height: 1564,
-        },
-        "973": {
-            advance: 1065,
-            bearingX: 74,
-            bearingY: 1505,
-            width: 923,
-            height: 1567,
-        },
-        "974": {
-            advance: 1557,
-            bearingX: 56,
-            bearingY: 1505,
-            width: 1398,
-            height: 1547,
-        },
-        "1025": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 1948,
-            width: 1073,
-            height: 2047,
-        },
-        "1026": {
-            advance: 1741,
-            bearingX: 42,
-            bearingY: 1515,
-            width: 1600,
-            height: 1794,
-        },
-        "1027": {
-            advance: 1262,
-            bearingX: 182,
-            bearingY: 2115,
-            width: 1007,
-            height: 2123,
-        },
-        "1028": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 1523,
-            width: 1113,
-            height: 1547,
-        },
-        "1029": {
-            advance: 1420,
-            bearingX: 133,
-            bearingY: 1467,
-            width: 1188,
-            height: 1524,
-        },
-        "1030": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1494,
-            width: 985,
-            height: 1530,
-        },
-        "1031": {
-            advance: 1119,
-            bearingX: 75,
-            bearingY: 1850,
-            width: 985,
-            height: 1886,
-        },
-        "1032": {
-            advance: 1362,
-            bearingX: 94,
-            bearingY: 1515,
-            width: 1202,
-            height: 1635,
-        },
-        "1033": {
-            advance: 2048,
-            bearingX: 25,
-            bearingY: 1525,
-            width: 1954,
-            height: 1544,
-        },
-        "1034": {
-            advance: 2001,
-            bearingX: 48,
-            bearingY: 1553,
-            width: 1872,
-            height: 1579,
-        },
-        "1035": {
-            advance: 1734,
-            bearingX: 42,
-            bearingY: 1518,
-            width: 1544,
-            height: 1558,
-        },
-        "1036": {
-            advance: 1297,
-            bearingX: 161,
-            bearingY: 2115,
-            width: 1051,
-            height: 2131,
-        },
-        "1038": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 2039,
-            width: 1196,
-            height: 2077,
-        },
-        "1039": {
-            advance: 1494,
-            bearingX: 123,
-            bearingY: 1536,
-            width: 1262,
-            height: 1814,
-        },
-        "1040": {
-            advance: 1498,
-            bearingX: 131,
-            bearingY: 1478,
-            width: 1221,
-            height: 1508,
-        },
-        "1041": {
-            advance: 1238,
-            bearingX: 155,
-            bearingY: 1522,
-            width: 975,
-            height: 1536,
-        },
-        "1042": {
-            advance: 1291,
-            bearingX: 191,
-            bearingY: 1571,
-            width: 1017,
-            height: 1619,
-        },
-        "1043": {
-            advance: 1262,
-            bearingX: 182,
-            bearingY: 1515,
-            width: 1007,
-            height: 1523,
-        },
-        "1044": {
-            advance: 1573,
-            bearingX: 90,
-            bearingY: 1543,
-            width: 1400,
-            height: 1877,
-        },
-        "1045": {
-            advance: 1279,
-            bearingX: 140,
-            bearingY: 1605,
-            width: 1073,
-            height: 1704,
-        },
-        "1046": {
-            advance: 1760,
-            bearingX: 30,
-            bearingY: 1538,
-            width: 1688,
-            height: 1554,
-        },
-        "1047": {
-            advance: 1231,
-            bearingX: 66,
-            bearingY: 1508,
-            width: 1055,
-            height: 1550,
-        },
-        "1048": {
-            advance: 1531,
-            bearingX: 129,
-            bearingY: 1548,
-            width: 1259,
-            height: 1600,
-        },
-        "1049": {
-            advance: 1531,
-            bearingX: 129,
-            bearingY: 2039,
-            width: 1259,
-            height: 2091,
-        },
-        "1050": {
-            advance: 1297,
-            bearingX: 161,
-            bearingY: 1538,
-            width: 1051,
-            height: 1554,
-        },
-        "1051": {
-            advance: 1513,
-            bearingX: 25,
-            bearingY: 1525,
-            width: 1310,
-            height: 1547,
-        },
-        "1052": {
-            advance: 1808,
-            bearingX: 112,
-            bearingY: 1534,
-            width: 1619,
-            height: 1617,
-        },
-        "1053": {
-            advance: 1573,
-            bearingX: 152,
-            bearingY: 1554,
-            width: 1313,
-            height: 1637,
-        },
-        "1054": {
-            advance: 1635,
-            bearingX: 116,
-            bearingY: 1515,
-            width: 1432,
-            height: 1576,
-        },
-        "1055": {
-            advance: 1860,
-            bearingX: 105,
-            bearingY: 1568,
-            width: 1651,
-            height: 1605,
-        },
-        "1056": {
-            advance: 1066,
-            bearingX: 100,
-            bearingY: 1571,
-            width: 905,
-            height: 1595,
-        },
-        "1057": {
-            advance: 1234,
-            bearingX: 90,
-            bearingY: 1523,
-            width: 1113,
-            height: 1547,
-        },
-        "1058": {
-            advance: 1392,
-            bearingX: 116,
-            bearingY: 1518,
-            width: 1347,
-            height: 1526,
-        },
-        "1059": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 1521,
-            width: 1196,
-            height: 1559,
-        },
-        "1060": {
-            advance: 1307,
-            bearingX: 91,
-            bearingY: 1494,
-            width: 1181,
-            height: 1530,
-        },
-        "1061": {
-            advance: 1482,
-            bearingX: 68,
-            bearingY: 1521,
-            width: 1338,
-            height: 1605,
-        },
-        "1062": {
-            advance: 1584,
-            bearingX: 123,
-            bearingY: 1536,
-            width: 1413,
-            height: 1832,
-        },
-        "1063": {
-            advance: 1264,
-            bearingX: 105,
-            bearingY: 1544,
-            width: 1017,
-            height: 1574,
-        },
-        "1064": {
-            advance: 1856,
-            bearingX: 99,
-            bearingY: 1536,
-            width: 1637,
-            height: 1552,
-        },
-        "1065": {
-            advance: 1960,
-            bearingX: 99,
-            bearingY: 1536,
-            width: 1812,
-            height: 1832,
-        },
-        "1066": {
-            advance: 1556,
-            bearingX: 38,
-            bearingY: 1517,
-            width: 1410,
-            height: 1531,
-        },
-        "1067": {
-            advance: 2202,
-            bearingX: 155,
-            bearingY: 1514,
-            width: 1953,
-            height: 1550,
-        },
-        "1068": {
-            advance: 1238,
-            bearingX: 155,
-            bearingY: 1514,
-            width: 975,
-            height: 1528,
-        },
-        "1069": {
-            advance: 1340,
-            bearingX: 88,
-            bearingY: 1500,
-            width: 1147,
-            height: 1513,
-        },
-        "1070": {
-            advance: 2315,
-            bearingX: 81,
-            bearingY: 1537,
-            width: 2147,
-            height: 1598,
-        },
-        "1071": {
-            advance: 1287,
-            bearingX: 26,
-            bearingY: 1541,
-            width: 1148,
-            height: 1578,
-        },
-        "1072": {
-            advance: 1048,
-            bearingX: 51,
-            bearingY: 1044,
-            width: 962,
-            height: 1110,
-        },
-        "1073": {
-            advance: 1077,
-            bearingX: 76,
-            bearingY: 1580,
-            width: 956,
-            height: 1639,
-        },
-        "1074": {
-            advance: 1032,
-            bearingX: 131,
-            bearingY: 1048,
-            width: 827,
-            height: 1094,
-        },
-        "1075": {
-            advance: 963,
-            bearingX: 134,
-            bearingY: 1059,
-            width: 803,
-            height: 1105,
-        },
-        "1076": {
-            advance: 1214,
-            bearingX: 33,
-            bearingY: 1063,
-            width: 1134,
-            height: 1325,
-        },
-        "1077": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1045,
-            width: 996,
-            height: 1091,
-        },
-        "1078": {
-            advance: 1398,
-            bearingX: 66,
-            bearingY: 1061,
-            width: 1234,
-            height: 1087,
-        },
-        "1079": {
-            advance: 900,
-            bearingX: 106,
-            bearingY: 1071,
-            width: 688,
-            height: 1118,
-        },
-        "1080": {
-            advance: 1304,
-            bearingX: 99,
-            bearingY: 1050,
-            width: 1045,
-            height: 1108,
-        },
-        "1081": {
-            advance: 1304,
-            bearingX: 99,
-            bearingY: 1589,
-            width: 1045,
-            height: 1647,
-        },
-        "1082": {
-            advance: 1008,
-            bearingX: 129,
-            bearingY: 1061,
-            width: 840,
-            height: 1087,
-        },
-        "1083": {
-            advance: 1325,
-            bearingX: 30,
-            bearingY: 1064,
-            width: 1153,
-            height: 1103,
-        },
-        "1084": {
-            advance: 1504,
-            bearingX: 44,
-            bearingY: 1064,
-            width: 1400,
-            height: 1122,
-        },
-        "1085": {
-            advance: 1146,
-            bearingX: 75,
-            bearingY: 1107,
-            width: 950,
-            height: 1160,
-        },
-        "1086": {
-            advance: 1211,
-            bearingX: 69,
-            bearingY: 1086,
-            width: 1075,
-            height: 1148,
-        },
-        "1087": {
-            advance: 1325,
-            bearingX: 212,
-            bearingY: 1046,
-            width: 975,
-            height: 1090,
-        },
-        "1088": {
-            advance: 1095,
-            bearingX: 119,
-            bearingY: 1099,
-            width: 889,
-            height: 1680,
-        },
-        "1089": {
-            advance: 1052,
-            bearingX: 105,
-            bearingY: 1063,
-            width: 864,
-            height: 1126,
-        },
-        "1090": {
-            advance: 965,
-            bearingX: 65,
-            bearingY: 1057,
-            width: 841,
-            height: 1121,
-        },
-        "1091": {
-            advance: 1066,
-            bearingX: -4,
-            bearingY: 1040,
-            width: 1028,
-            height: 1619,
-        },
-        "1092": {
-            advance: 1659,
-            bearingX: 85,
-            bearingY: 1099,
-            width: 1518,
-            height: 1680,
-        },
-        "1093": {
-            advance: 1209,
-            bearingX: 61,
-            bearingY: 1105,
-            width: 1090,
-            height: 1149,
-        },
-        "1094": {
-            advance: 1267,
-            bearingX: 177,
-            bearingY: 1084,
-            width: 1032,
-            height: 1335,
-        },
-        "1095": {
-            advance: 1093,
-            bearingX: 75,
-            bearingY: 1090,
-            width: 907,
-            height: 1142,
-        },
-        "1096": {
-            advance: 1581,
-            bearingX: 107,
-            bearingY: 1084,
-            width: 1368,
-            height: 1135,
-        },
-        "1097": {
-            advance: 1624,
-            bearingX: 107,
-            bearingY: 1084,
-            width: 1498,
-            height: 1417,
-        },
-        "1098": {
-            advance: 1272,
-            bearingX: 17,
-            bearingY: 1089,
-            width: 1202,
-            height: 1103,
-        },
-        "1099": {
-            advance: 1619,
-            bearingX: 128,
-            bearingY: 1056,
-            width: 1437,
-            height: 1101,
-        },
-        "1100": {
-            advance: 1007,
-            bearingX: 128,
-            bearingY: 1056,
-            width: 791,
-            height: 1070,
-        },
-        "1101": {
-            advance: 1051,
-            bearingX: 108,
-            bearingY: 1054,
-            width: 852,
-            height: 1095,
-        },
-        "1102": {
-            advance: 1848,
-            bearingX: 84,
-            bearingY: 1121,
-            width: 1677,
-            height: 1183,
-        },
-        "1103": {
-            advance: 1035,
-            bearingX: 34,
-            bearingY: 1048,
-            width: 880,
-            height: 1085,
-        },
-        "1105": {
-            advance: 1122,
-            bearingX: 87,
-            bearingY: 1477,
-            width: 996,
-            height: 1523,
-        },
-        "1106": {
-            advance: 1147,
-            bearingX: -67,
-            bearingY: 1602,
-            width: 1089,
-            height: 2011,
-        },
-        "1107": {
-            advance: 963,
-            bearingX: 134,
-            bearingY: 1665,
-            width: 803,
-            height: 1711,
-        },
-        "1108": {
-            advance: 1051,
-            bearingX: 79,
-            bearingY: 1046,
-            width: 852,
-            height: 1096,
-        },
-        "1109": {
-            advance: 997,
-            bearingX: 41,
-            bearingY: 1142,
-            width: 871,
-            height: 1202,
-        },
-        "1110": {
-            advance: 574,
-            bearingX: 179,
-            bearingY: 1498,
-            width: 268,
-            height: 1503,
-        },
-        "1111": {
-            advance: 574,
-            bearingX: -122,
-            bearingY: 1380,
-            width: 754,
-            height: 1385,
-        },
-        "1112": {
-            advance: 826,
-            bearingX: -18,
-            bearingY: 1496,
-            width: 674,
-            height: 2093,
-        },
-        "1113": {
-            advance: 1822,
-            bearingX: 30,
-            bearingY: 1083,
-            width: 1743,
-            height: 1121,
-        },
-        "1114": {
-            advance: 1585,
-            bearingX: 89,
-            bearingY: 1057,
-            width: 1413,
-            height: 1079,
-        },
-        "1115": {
-            advance: 1183,
-            bearingX: -67,
-            bearingY: 1602,
-            width: 1145,
-            height: 1664,
-        },
-        "1116": {
-            advance: 1008,
-            bearingX: 129,
-            bearingY: 1665,
-            width: 840,
-            height: 1691,
-        },
-        "1118": {
-            advance: 1066,
-            bearingX: -4,
-            bearingY: 1589,
-            width: 1028,
-            height: 2168,
-        },
-        "1119": {
-            advance: 1267,
-            bearingX: 137,
-            bearingY: 1084,
-            width: 958,
-            height: 1427,
-        },
-        "1168": {
-            advance: 1262,
-            bearingX: 182,
-            bearingY: 1777,
-            width: 1007,
-            height: 1785,
-        },
-        "1169": {
-            advance: 963,
-            bearingX: 134,
-            bearingY: 1342,
-            width: 809,
-            height: 1388,
-        },
-        "7808": {
-            advance: 2129,
-            bearingX: 139,
-            bearingY: 2111,
-            width: 1957,
-            height: 2207,
-        },
-        "7809": {
-            advance: 1401,
-            bearingX: 76,
-            bearingY: 1661,
-            width: 1270,
-            height: 1743,
-        },
-        "7810": {
-            advance: 2129,
-            bearingX: 139,
-            bearingY: 2115,
-            width: 1957,
-            height: 2211,
-        },
-        "7811": {
-            advance: 1401,
-            bearingX: 76,
-            bearingY: 1665,
-            width: 1270,
-            height: 1747,
-        },
-        "7812": {
-            advance: 2129,
-            bearingX: 139,
-            bearingY: 1830,
-            width: 1957,
-            height: 1926,
-        },
-        "7813": {
-            advance: 1401,
-            bearingX: 76,
-            bearingY: 1380,
-            width: 1270,
-            height: 1462,
-        },
-        "7922": {
-            advance: 1301,
-            bearingX: 29,
-            bearingY: 2111,
-            width: 1196,
-            height: 2181,
-        },
-        "7923": {
-            advance: 1066,
-            bearingX: -4,
-            bearingY: 1661,
-            width: 1028,
-            height: 2240,
-        },
-        "8211": {
-            advance: 904,
-            bearingX: 44,
-            bearingY: 759,
-            width: 784,
-            height: 215,
-        },
-        "8212": {
-            advance: 1808,
-            bearingX: 67,
-            bearingY: 748,
-            width: 1641,
-            height: 203,
-        },
-        "8213": {
-            advance: 1808,
-            bearingX: 67,
-            bearingY: 748,
-            width: 1641,
-            height: 203,
-        },
-        "8215": {
-            advance: 1284,
-            bearingX: -35,
-            bearingY: -57,
-            width: 1358,
-            height: 488,
-        },
-        "8216": {
-            advance: 369,
-            bearingX: 77,
-            bearingY: 1587,
-            width: 206,
-            height: 493,
-        },
-        "8217": {
-            advance: 369,
-            bearingX: 57,
-            bearingY: 1584,
-            width: 221,
-            height: 490,
-        },
-        "8218": {
-            advance: 612,
-            bearingX: 81,
-            bearingY: 224,
-            width: 325,
-            height: 418,
-        },
-        "8220": {
-            advance: 806,
-            bearingX: 71,
-            bearingY: 1615,
-            width: 597,
-            height: 526,
-        },
-        "8221": {
-            advance: 806,
-            bearingX: 104,
-            bearingY: 1580,
-            width: 623,
-            height: 582,
-        },
-        "8222": {
-            advance: 843,
-            bearingX: 81,
-            bearingY: 224,
-            width: 680,
-            height: 418,
-        },
-        "8224": {
-            advance: 1222,
-            bearingX: 51,
-            bearingY: 1697,
-            width: 1106,
-            height: 1905,
-        },
-        "8225": {
-            advance: 1222,
-            bearingX: 58,
-            bearingY: 1697,
-            width: 1107,
-            height: 1905,
-        },
-        "8226": {
-            advance: 793,
-            bearingX: 183,
-            bearingY: 959,
-            width: 400,
-            height: 401,
-        },
-        "8230": {
-            advance: 1383,
-            bearingX: 178,
-            bearingY: 161,
-            width: 1064,
-            height: 228,
-        },
-        "8240": {
-            advance: 2542,
-            bearingX: 147,
-            bearingY: 1642,
-            width: 2223,
-            height: 1672,
-        },
-        "8242": {
-            advance: 369,
-            bearingX: 57,
-            bearingY: 1584,
-            width: 221,
-            height: 490,
-        },
-        "8243": {
-            advance: 678,
-            bearingX: 57,
-            bearingY: 1584,
-            width: 561,
-            height: 490,
-        },
-        "8249": {
-            advance: 781,
-            bearingX: 106,
-            bearingY: 1061,
-            width: 608,
-            height: 867,
-        },
-        "8250": {
-            advance: 781,
-            bearingX: 106,
-            bearingY: 1061,
-            width: 608,
-            height: 867,
-        },
-        "8252": {
-            advance: 921,
-            bearingX: 120,
-            bearingY: 1605,
-            width: 652,
-            height: 1671,
-        },
-        "8254": {
-            advance: 1250,
-            bearingX: -154,
-            bearingY: 1626,
-            width: 1471,
-            height: 200,
-        },
-        "8260": {
-            advance: 1661,
-            bearingX: 26,
-            bearingY: 1700,
-            width: 1576,
-            height: 1754,
-        },
-        "8319": {
-            advance: 1079,
-            bearingX: 195,
-            bearingY: 1599,
-            width: 675,
-            height: 798,
-        },
-        "8355": {
-            advance: 2026,
-            bearingX: 173,
-            bearingY: 1573,
-            width: 1787,
-            height: 1679,
-        },
-        "8356": {
-            advance: 1625,
-            bearingX: 84,
-            bearingY: 1610,
-            width: 1372,
-            height: 1760,
-        },
-        "8359": {
-            advance: 2435,
-            bearingX: 100,
-            bearingY: 1571,
-            width: 2239,
-            height: 1635,
-        },
-        "8364": {
-            advance: 1250,
-            bearingX: -45,
-            bearingY: 1583,
-            width: 1292,
-            height: 1617,
-        },
-        "8453": {
-            advance: 1680,
-            bearingX: 78,
-            bearingY: 1603,
-            width: 1531,
-            height: 1633,
-        },
-        "8467": {
-            advance: 1001,
-            bearingX: 20,
-            bearingY: 1640,
-            width: 812,
-            height: 1681,
-        },
-        "8470": {
-            advance: 2370,
-            bearingX: 123,
-            bearingY: 1551,
-            width: 2129,
-            height: 1630,
-        },
-        "8480": {
-            advance: 2384,
-            bearingX: 41,
-            bearingY: 1469,
-            width: 2283,
-            height: 1124,
-        },
-        "8482": {
-            advance: 1808,
-            bearingX: 77,
-            bearingY: 1609,
-            width: 1668,
-            height: 763,
-        },
-        "8486": {
-            advance: 1966,
-            bearingX: 65,
-            bearingY: 1630,
-            width: 1836,
-            height: 1770,
-        },
-        "8494": {
-            advance: 1125,
-            bearingX: 61,
-            bearingY: 1091,
-            width: 1002,
-            height: 1111,
-        },
-        "8539": {
-            advance: 1333,
-            bearingX: 162,
-            bearingY: 1724,
-            width: 1049,
-            height: 2114,
-        },
-        "8540": {
-            advance: 1333,
-            bearingX: 162,
-            bearingY: 1701,
-            width: 1049,
-            height: 2091,
-        },
-        "8541": {
-            advance: 1333,
-            bearingX: 162,
-            bearingY: 1715,
-            width: 1049,
-            height: 2105,
-        },
-        "8542": {
-            advance: 1333,
-            bearingX: 162,
-            bearingY: 1711,
-            width: 1049,
-            height: 2101,
-        },
-        "8706": {
-            advance: 1042,
-            bearingX: 185,
-            bearingY: 1538,
-            width: 806,
-            height: 1579,
-        },
-        "8710": {
-            advance: 1627,
-            bearingX: 98,
-            bearingY: 1663,
-            width: 1503,
-            height: 1735,
-        },
-        "8719": {
-            advance: 1860,
-            bearingX: 105,
-            bearingY: 1568,
-            width: 1651,
-            height: 1845,
-        },
-        "8721": {
-            advance: 1530,
-            bearingX: 45,
-            bearingY: 1486,
-            width: 1411,
-            height: 1799,
-        },
-        "8722": {
-            advance: 984,
-            bearingX: 48,
-            bearingY: 737,
-            width: 865,
-            height: 191,
-        },
-        "8730": {
-            advance: 1244,
-            bearingX: 50,
-            bearingY: 1630,
-            width: 1153,
-            height: 1699,
-        },
-        "8734": {
-            advance: 1714,
-            bearingX: 147,
-            bearingY: 962,
-            width: 1418,
-            height: 656,
-        },
-        "8747": {
-            advance: 1433,
-            bearingX: 80,
-            bearingY: 1720,
-            width: 1233,
-            height: 2124,
-        },
-        "8776": {
-            advance: 1269,
-            bearingX: 123,
-            bearingY: 1199,
-            width: 1082,
-            height: 953,
-        },
-        "8800": {
-            advance: 1045,
-            bearingX: 99,
-            bearingY: 1224,
-            width: 781,
-            height: 1141,
-        },
-        "8804": {
-            advance: 781,
-            bearingX: 19,
-            bearingY: 1061,
-            width: 735,
-            height: 1189,
-        },
-        "8805": {
-            advance: 781,
-            bearingX: 35,
-            bearingY: 1125,
-            width: 719,
-            height: 1253,
-        },
-        "9472": {
-            advance: 1451,
-            bearingX: -23,
-            bearingY: 709,
-            width: 1496,
-            height: 175,
-        },
-        "9474": {
-            advance: 1451,
-            bearingX: 638,
-            bearingY: 1864,
-            width: 175,
-            height: 2485,
-        },
-        "9484": {
-            advance: 1451,
-            bearingX: 638,
-            bearingY: 709,
-            width: 836,
-            height: 1330,
-        },
-        "9488": {
-            advance: 1451,
-            bearingX: -23,
-            bearingY: 709,
-            width: 835,
-            height: 1330,
-        },
-        "9492": {
-            advance: 1451,
-            bearingX: 638,
-            bearingY: 1864,
-            width: 836,
-            height: 1330,
-        },
-        "9496": {
-            advance: 1451,
-            bearingX: -23,
-            bearingY: 1864,
-            width: 835,
-            height: 1330,
-        },
-        "9633": {
-            advance: 1504,
-            bearingX: 364,
-            bearingY: 780,
-            width: 779,
-            height: 780,
-        },
-        "9642": {
-            advance: 1316,
-            bearingX: 386,
-            bearingY: 656,
-            width: 532,
-            height: 532,
-        },
-        "9643": {
-            advance: 1316,
-            bearingX: 369,
-            bearingY: 674,
-            width: 567,
-            height: 568,
-        },
-        "9674": {
-            advance: 1479,
-            bearingX: 60,
-            bearingY: 1635,
-            width: 1370,
-            height: 1717,
-        },
-        "9679": {
-            advance: 1316,
-            bearingX: 376,
-            bearingY: 663,
-            width: 542,
-            height: 542,
-        },
-        "9702": {
-            advance: 1316,
-            bearingX: 358,
-            bearingY: 681,
-            width: 578,
-            height: 578,
-        },
-        "9834": {
-            advance: 1024,
-            bearingX: 37,
-            bearingY: 1363,
-            width: 950,
-            height: 1400,
-        },
-        "64257": {
-            advance: 1349,
-            bearingX: 75,
-            bearingY: 1599,
-            width: 1147,
-            height: 1759,
-        },
-        "64258": {
-            advance: 1343,
-            bearingX: 70,
-            bearingY: 1617,
-            width: 1168,
-            height: 1777,
-        },
-    },
-};
-
-export default metrics;
