+++ packages/metrics/src/font-metrics-context.ts
@@ -1,0 +1,15 @@
+import * as React from "react";
+
+import type {FontMetrics} from "./types";
+
+const placeholderMetrics: FontMetrics = {
+    unitsPerEm: 1000,
+    ascender: 850,
+    descender: 150,
+    getGlyphMetrics: (charCode: number) => null,
+    glyphMetrics: {},
+};
+
+export const FontMetricsContext = React.createContext<FontMetrics>(
+    placeholderMetrics,
+);
