+++ packages/typesetter/package.json
@@ -4,6 +4,7 @@
     "main": "./out/src/index.js",
     "dependencies": {
         "@math-blocks/core": "0.0.1",
-        "@math-blocks/editor-core": "0.0.1"
+        "@math-blocks/editor-core": "0.0.1",
+        "@math-blocks/metrics": "0.0.1"
     }
 }
