--- packages/typesetter/src/metrics.ts
@@ -1,16 +1,0 @@
-export type GlyphMetrics = {
-    advance: number;
-    bearingX: number;
-    bearingY: number;
-    width: number;
-    height: number;
-};
-
-export type FontMetrics = {
-    unitsPerEm: number;
-    ascender: number;
-    descender: number;
-    glyphMetrics: Partial<{
-        [charCode: number]: GlyphMetrics;
-    }>;
-};
