+++ packages/typesetter/tsconfig.json
@@ -7,6 +7,7 @@
   },
   "references": [
     {"path": "../core"},
-    {"path": "../editor-core"}
+    {"path": "../editor-core"},
+    {"path": "../metrics"}
   ]
 }
